requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.5.2
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docopt==0.6.2
docstring_parser==0.16
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
google-adk==0.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.2
google-cloud-speech==2.31.1
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.10.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
graphviz==0.20.3
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
instagrapi==2.1.3
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.6.0
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.4
opentelemetry-api==1.32.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic-settings==2.8.1
pydantic_core==2.27.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
yarg==0.1.9
zipp==3.21.0
